--- tox.ini
@@ -1,22 +1,0 @@
-[tox]
-envlist = py35, py36, flake8
-
-[testenv:flake8]
-basepython=python
-deps=flake8
-commands=flake8 mychevy
-
-[testenv]
-setenv =
-    PYTHONPATH = {toxinidir}
-deps =
-    -r{toxinidir}/requirements_dev.txt
-commands =
-    pip install -U pip
-    py.test --basetemp={envtmpdir}
-
-
-; If you want to make tox run the tests with the same versions, create a
-; requirements.txt with the pinned versions and uncomment the following lines:
-; deps =
-;     -r{toxinidir}/requirements.txt
